Allow watching additional directories
Problem: I'm working on a Go application with html/template files and I
need to manually restart the binary every time to see any change.

Solution: allow giving an additional list of directories to watch. For example:

	var tpl = loadTpl()

	func loadTpl() *template.Template {
		t, err := template.New("").Option("missingkey=error").ParseGlob("tpl/*.gohtml")
		if err != nil {
			log.Println(err)
		}
		return t
	}

	func main() {
		go func() {
			err := reload.Do(log.Printf, reload.Dir("./tpl", func() {
				log.Printf("reloading templates")
				tpl = loadTpl()
			}))
		}()
	}

Which seems to work pretty well. Great success!

This should be a backwards-compatible change.